nemo_terminal.py: Change how we check for a busy terminal to checking for a child process.
This is a much more stable method to use that covers nearly all cases.
Checking wchan was prone to breaking every so often as kernel changes
would alter how wchan behaves.  Simply checking to see if our pseudo-
term process has a child is the same method gnome-terminal uses to
make sure a window isn't busy before closing it, and, it turns out,
is the same way nautilus-terminal does this check as well (though in
a slightly different manner.)

The only thing this would fail to catch is something like a 'read'
in a script, which isn't a program, but a built-in command, but I
think the likelyhood of this being a problem in the context of a
nemo extension is pretty low.

Fixes #134