install-samba: prevent unprivileged users to add other users to sambashare
The way install-samba is currently implemented the command line argument
is prefered to the PKEXEC_UID environment variable. This means if an
admin relaxes the polkit policy to allow non-privileged users to run
this script without admin password, then non-privileged users could add
arbitrary other users to the sambashare group.

This is certainly not the purpose of this helper script. Therefore the
PKEXEC_UID should be preferred and the command line arguments be ignored
if it is present.